chore: sync to main branch (#23)
* feat: add notification

* fix: add missing user to reviews, comments for tests

* feat: add like, mention notifications for comments

* feat: add like notifications for reviews

* feat: add user coin added notifications

* feat: add list notifications

* chore: update notification service

* remove title column & add computed title attribute

* feat: add announcement notiffications

* feat: add reviews notifications

* feat: add chapters notifications

* add followers state to novel factory

* feat: add comments notifications tests

* fix: update min value to limit for promocode factory

* chore: number column int to decimal

* chore: improve validation messages to use promocode request

* fix: add missing coin_amount, free_amount attributes for novel

* chore: update on coin added notifications

* fix: update comment notifications

* fix: update warning message to use promocode

* chore: sort tags

* fix: wrong startDate to reports

* fix: wrong username to reviews, comments tests

* chore: add novel, chapter href to notification service

* refactor: update body to notification service

* chore: add missing on delete if destroy for notifications

* chore: update announcement notifications

* chore: update validation messages to purchase chapter

* feat: add dependabot

* feat: add ci

* feat: add muted ability on comments (#7)

* fix: extract cloudflare r2 url to env (#4)

* feat: add admin unmute endpoint

* tests: update user unmute tests

* fix: update admin unmute endpoint

* feat: add public novel image endpoint

* feat: add author to og image endpoit

* chore(deps-dev): bump eslint from 8.21.0 to 8.30.0 (#19)


* chore(deps-dev): bump typescript from 4.6.4 to 4.9.4 (#9)


* chore(deps-dev): bump prettier from 2.7.1 to 2.8.1 (#10)


* fix: update register request regexes

* refactor: combine validate message to use promocode

* fix: wrong notificationable type for like review

* chore: add fields prop to on delete for notification service

* refactor: like comment, review & reaction service & tests

* chore: add missing auth middleware

* fix: wrong chapter url

* chore: update full name to chapter

* chore: update body to new chapter notification

* fix: wrong notification of liking own comment

* fix: null href to new chapter notification

* fix: wrong announcement url

* Fix/password regex (#22)

* fix: update password regex

* chore: remove unused update profile endpoint

* chore: update user validator

* chore: password required if update email

* feat: add notifications to user

* fix: wrong dislike reaction to notification

+++ database/migrations/1664564378068_volumes.ts
@@ -12,7 +12,7 @@ export default class extends BaseSchema {
       table.foreign('volume_novel_id').references('novels.id').onDelete('CASCADE')
 
       table.string('name').nullable()
-      table.integer('volume_number')
+      table.decimal('volume_number')
 
       table.enu('publish_status', Object.values(VolumePublishStatus))
